\name{biosyntax-package}
\alias{biosyntax-package}
\docType{package}
\title{Biological syntax highlighting for R}
\description{
The biosyntax package highlights biological text as ANSI, HTML, or semantic
spans. It also accepts supported Bioconductor objects.
}
\details{
The package embeds libbiosyntax. It highlights text but does not parse or alter
biological data.
}
\keyword{package}
